+++ ceammc/CHANGELOG.md
@@ -5,6 +5,8 @@
 - new properties:
   - @norm property added to ui.env to do envelope value normalization on input
   - @clip property added to system.cursor to clip cursor XY coords
+- new methods:
+  - [polltime( method added to system.cursor to change global polltime
 
 ### Fixed:
 - system.cursor no normalization in subpatch